Auto-publish the spec to gh-pages
This update introduces a GitHub job that uses the w3c/spec-prod action to
build the spec with Bikeshed, run checks, and publish the result to the
`gh-pages` branch.

That job will also take care of publishing the spec to `https://www.w3.org/TR`
once it has been published as First Public Working Draft (until then, the
`ECHIDNA_TOKEN` will not be set and that step won't happen).

This update also enables PR Preview to be able to visualize the HTML diff in
pull requests, and tells GitHub to highlight the Bikeshed source as HTML.