Adjust overloaded members to please Bikeshed
Bikeshed does not like it when overloaded methods have the exact same argument
names. Now using different names.

This update also makes the "Repository" setting explicit to generate the right
GitHub links in the spec's front matter (the published document currently points
to a fork). This also allows to generate the document locally outside of a
GitHub repo, or through Bikeshed's Web form.

Bikeshed also reported a warning because it knows of two possible specs that
have a `document` IDL. It selects the right reference in the HTML spec, but this
update makes that explicit to avoid any possible issue in the future.