[meta] Enable auto-publish to TR
This update adds needed parameters to the auto-publish job for it to take care
of publication to https://www.w3.org/TR/ as well.

It also updates the spec metadata to add the missing w3cid for the editor, the
/TR URL, and drop the level to stick to an `autoplay-detection` shortname for
now. This can be changed later on if group decides that a level-based approach
is needed.

(Note that the Echidna token referenced in the job was added to the repository
settings as secret)